Expiry notices: move shared banner helpers to the loader
